+++ caf-propertysource-spring/src/main/java/com/github/cafapi/util/spring/propertysource/CafConfigPropertySource.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafapi.util.spring;
+package com.github.cafapi.util.spring.propertysource;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
